io.quarkus.it package fixes
