+++ integration-tests/keycloak/src/main/java/io/quarkus/it/keycloak/TestConfigResolver.java
@@ -1,4 +1,4 @@
-package io.quarkus.example.keycloak;
+package io.quarkus.it.keycloak;
 
 import java.util.HashMap;
 import java.util.Map;
