+++ integration-tests/main/src/test/java/io/quarkus/it/main/CustomConfigSourceITCase.java
@@ -1,4 +1,4 @@
-package io.quarkus.example.test;
+package io.quarkus.it.main;
 
 import io.quarkus.test.junit.SubstrateTest;
